refactor(packet): split modules into separate src files
