+++ src/tracing.rs
@@ -1,11 +1,13 @@
 mod config;
 mod error;
 mod net;
-mod packet;
 mod probe;
 mod tracer;
 mod types;
 mod util;
+
+/// Packet wire formats.
+pub mod packet;
 
 pub use config::{
     PortDirection, TracerAddrFamily, TracerChannelConfig, TracerConfig, TracerProtocol,
@@ -13,12 +15,3 @@ pub use config::{
 pub use net::TracerChannel;
 pub use probe::{IcmpPacketType, Probe, ProbeStatus};
 pub use tracer::{Tracer, TracerRound};
-
-pub use packet::icmp::{IcmpCode, IcmpPacket, IcmpType};
-
-pub use packet::icmp::destination_unreachable::DestinationUnreachablePacket;
-pub use packet::icmp::echo_reply::EchoReplyPacket;
-pub use packet::icmp::echo_request::EchoRequestPacket;
-pub use packet::icmp::time_exceeded::TimeExceededPacket;
-pub use packet::ipv4::Ipv4Packet;
-pub use packet::udp::UdpPacket;
